# docs/swagger.yaml
definitions:
...
  handler.ServiceRequest:
    properties:
      customer_id:
        type: integer
      payment_method:
        type: string
      services:
        items:
          properties:
            quantity:
              type: integer
            service_id:
              type: integer
          type: object
        type: array
    type: object
...
info:
  contact: {}
paths:
...
  /services/purchase:
    post:
      consumes:
      - application/json
      description: Allows customers to purchase services using either wallet or GoPay
        as the payment method. The endpoint validates admin roles.
      parameters:
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ServiceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Services purchased successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Purchase services
      tags:
      - Services
...
swagger: "2.0"